Follow track fixes + NSUInteger negative overflow fixes

+++ Sources/Controllers/2.0/BottomSheet/RoutePlanningBottomSheet/OAPointOptionsBottomSheetViewController.mm
@@ -254,7 +254,7 @@
                 cell.textColorNormal = nil;
                 cell.iconView.image = [UIImage imageNamed:item[@"img"]];
                 cell.titleView.textColor = UIColor.blackColor;
-                cell.separatorView.hidden = indexPath.row == _data[indexPath.section].count - 1;
+                cell.separatorView.hidden = indexPath.row == (NSInteger) _data[indexPath.section].count - 1;
             }
         }
         return cell;
